+++ app/account/ManageSubscriptionButton.tsx
@@ -13,7 +13,8 @@ export default function ManageSubscriptionButton({ session }) {
     } catch (error) {
       if (error) return alert((error as Error).message);
     }
-  }; //experimental server action
+  };
+
   return (
     <div className="flex flex-col items-start justify-between sm:flex-row sm:items-center">
       <p className="pb-4 sm:pb-0">Manage your subscription on Stripe.</p>
