+++ package.json
@@ -16,7 +16,7 @@
     "@supabase/auth-ui-shared": "^0.1.6",
     "@supabase/supabase-js": "^2.21.0",
     "classnames": "^2.3.2",
-    "next": "^13.4.1",
+    "next": "13.4.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-merge-refs": "^2.0.2",
