+++ utils/supabase-client.ts
@@ -15,3 +15,4 @@ export const updateUserName = async (user: User, name: string) => {
     })
     .eq('id', user.id);
 };
+
