+++ utils/supabase-server.ts
@@ -7,3 +7,40 @@ export const createServerSupabaseClient = () =>
     headers,
     cookies
   });
+
+export async function getUserData() {
+  const supabase = createServerSupabaseClient();
+  try {
+    const {
+      data: { session }
+    } = await supabase.auth.getSession();
+    const {
+      data: { user }
+    } = await supabase.auth.getUser();
+    const { data: userDetails } = await supabase
+      .from('users')
+      .select('*')
+      .single();
+    const { data: subscription } = await supabase
+      .from('subscriptions')
+      .select('*, prices(*, products(*))')
+      .in('status', ['trialing', 'active'])
+      .single();
+
+    return {
+      session,
+      user,
+      userDetails,
+      subscription,
+      isLoading: false
+    };
+  } catch (error) {
+    console.error('Error:', error);
+    return {
+      session: null,
+      user: null,
+      userDetails: null,
+      subscription: null
+    };
+  }
+}
